chore: move `parse_catalog_and_schema_from_client_database_name` to catalog crate
