+++ src/servers/src/http/prom_store.rs
@@ -17,6 +17,7 @@ use axum::extract::{Query, RawBody, State};
 use axum::http::{header, StatusCode};
 use axum::response::IntoResponse;
 use common_catalog::consts::DEFAULT_SCHEMA_NAME;
+use common_catalog::parse_catalog_and_schema_from_db_string;
 use common_telemetry::timer;
 use hyper::Body;
 use prost::Message;
@@ -26,7 +27,6 @@ use session::context::QueryContext;
 use snafu::prelude::*;
 
 use crate::error::{self, Result};
-use crate::parse_catalog_and_schema_from_client_database_name;
 use crate::prom_store::snappy_decompress;
 use crate::query_handler::{PromStoreProtocolHandlerRef, PromStoreResponse};
 
@@ -59,7 +59,7 @@ pub async fn remote_write(
         )]
     );
     let ctx = if let Some(db) = params.db {
-        let (catalog, schema) = parse_catalog_and_schema_from_client_database_name(&db);
+        let (catalog, schema) = parse_catalog_and_schema_from_db_string(&db);
         QueryContext::with(catalog, schema)
     } else {
         QueryContext::arc()
@@ -99,7 +99,7 @@ pub async fn remote_read(
         )]
     );
     let ctx = if let Some(db) = params.db {
-        let (catalog, schema) = parse_catalog_and_schema_from_client_database_name(&db);
+        let (catalog, schema) = parse_catalog_and_schema_from_db_string(&db);
         QueryContext::with(catalog, schema)
     } else {
         QueryContext::arc()
